outline_LOC_Classification/topic.html now passes https://validator.w3.org/#validate_by_upload with "No errors or warnings to show.".
HTML5 considers formatting in the .html as obsolete and generates errors when validated. This follows the standard solution of moving the formatting to .css. See the new datatables_custom.css.

Multiple uses of an id in divs in one .html are not proper. The use of two #footer_text_right ids was fixed by creating #footer_text_right2 for the second one and using it in the .html. An alternate solution which might be better is to create and use a class. Classes can be used multiple times in a .html file.